+++ Robust.Shared/Physics/SharedBroadphaseSystem.cs
@@ -490,88 +490,6 @@ namespace Robust.Shared.Physics
             _queuedRotates.Enqueue(args);
         }
 
-        public void CreateFixture(PhysicsComponent body, Fixture fixture)
-        {
-            fixture.ID = body.GetFixtureName(fixture);
-            body.Fixtures.Add(fixture);
-            body.FixtureCount += 1;
-            fixture.Body = body;
-
-            // TODO: Assert world locked
-            // Broadphase should be set in the future TM
-            // Should only happen for nullspace / initializing entities
-            if (body.Broadphase != null)
-            {
-                UpdateBroadphaseCache(body.Broadphase);
-                CreateProxies(fixture, body.Owner.Transform.WorldPosition, false);
-            }
-
-            // Supposed to be wrapped in density but eh
-            body.ResetMassData();
-            body.Dirty();
-            // TODO: Set newcontacts to true.
-        }
-
-        public Fixture CreateFixture(PhysicsComponent body, IPhysShape shape)
-        {
-            var fixture = new Fixture(body, shape);
-            CreateFixture(body, fixture);
-            return fixture;
-        }
-
-        public void CreateFixture(PhysicsComponent body, IPhysShape shape, float mass)
-        {
-            // TODO: Make it take in density instead
-            var fixture = new Fixture(body, shape) {Mass = mass};
-            CreateFixture(body, fixture);
-        }
-
-        public void DestroyFixture(Fixture fixture)
-        {
-            DestroyFixture(fixture.Body, fixture);
-        }
-
-        public void DestroyFixture(PhysicsComponent body, Fixture fixture)
-        {
-            // TODO: Move this to fixturemanagersystem instead
-
-            // TODO: Assert world locked
-            DebugTools.Assert(fixture.Body == body);
-            DebugTools.Assert(body.FixtureCount > 0);
-
-            if (!body.Fixtures.Remove(fixture))
-            {
-                Logger.ErrorS("physics", $"Tried to remove fixture from {body.Owner} that was already removed.");
-                return;
-            }
-
-            var edge = body.ContactEdges;
-
-            while (edge != null)
-            {
-                var contact = edge.Contact!;
-                edge = edge.Next;
-
-                var fixtureA = contact.FixtureA;
-                var fixtureB = contact.FixtureB;
-
-                if (fixture == fixtureA || fixture == fixtureB)
-                {
-                    body.PhysicsMap?.ContactManager.Destroy(contact);
-                }
-            }
-
-            var broadphase = GetBroadphase(fixture.Body);
-
-            if (broadphase != null)
-            {
-                DestroyProxies(broadphase, fixture);
-            }
-
-            body.ResetMassData();
-            body.Dirty();
-        }
-
         private void SynchronizeFixtures(PhysicsComponent body, Vector2 worldPos, float worldRot)
         {
             // Logger.DebugS("physics", $"Synchronizing fixtures for {body.Owner}");
